Enforce bigint for all script numbers and bch amounts
- Use `bigint` rather than `number` for all instances of "script numbers"
  (e.g. function arguments) and satoshi amounts`
- Replace `contract.getRedeemScriptHex()` with `contract.bytecode`
- Update UTXO selection to use bigint (multiplied by 1e6 for "decimal points")

+++ packages/cashscript/src/network/BitcoinRpcNetworkProvider.ts
@@ -20,7 +20,7 @@ export default class BitcoinRpcNetworkProvider implements NetworkProvider {
     const utxos = result.map((utxo) => ({
       txid: utxo.txid,
       vout: utxo.vout,
-      satoshis: utxo.amount * 1e8,
+      satoshis: BigInt(utxo.amount * 1e8),
     }));
 
     return utxos;
